Refactor: move Key component to Keyboard.tsx
